Scrapping/nba/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.0
lxml>=4.9.0
sqlalchemy>=2.0.0
tqdm>=4.66.0
schedule>=1.2.0
loguru>=0.7.0
pyyaml>=6.0.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
